+++ app.js
@@ -3,7 +3,8 @@ const path = require("path");
 const bodyParser = require("body-parser");
 
 const shopRoutes = require("./routes/shop");
-const adminData = require("./routes/admin");
+const adminRoutes = require("./routes/admin");
+const errorController = require("./controllers/error");
 
 const app = express();
 
@@ -13,11 +14,9 @@ app.set("views", "views");
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(express.static(path.join(__dirname, "public")));
 
-app.use("/admin", adminData.routes);
+app.use("/admin", adminRoutes);
 app.use(shopRoutes);
 
-app.use((req, res, next) => {
-    res.render("404.ejs", { pageTitle: "Page Not Found", path: "" });
-});
+app.use(errorController.get404);
 
 app.listen(3001);
