+++ controllers/error.js
@@ -1,0 +1,3 @@
+exports.get404 = (req, res, next) => {
+    res.render("404.ejs", { pageTitle: "Page Not Found", path: "" });
+};
