+++ controllers/products.js
@@ -1,0 +1,27 @@
+const products = [];
+
+exports.getAddProduct = (req, res, next) => {
+    res.render("add-product.ejs", {
+        pageTitle: "Add Product",
+        path: "/admin/add-product",
+        activeAddProduct: true,
+        addProductCSS: true,
+        shopCSS: true,
+    });
+};
+
+exports.postAddProduct = (req, res, next) => {
+    products.push({ title: req.body.title });
+    res.redirect("/");
+};
+
+exports.getProducts = (req, res, next) => {
+    res.render("shop.ejs", {
+        prods: products,
+        pageTitle: "Shop",
+        path: "/",
+        hasProducts: products.length > 0,
+        activeShop: true,
+        shopCSS: true,
+    });
+};
