+++ routes/admin.js
@@ -1,26 +1,13 @@
 const express = require("express");
-const path = require("path");
+
+const productController = require("../controllers/products");
 
 const router = express.Router();
 
-const products = [];
-
 // /admin/add-product => POST
-router.post("/add-product", (req, res, next) => {
-    products.push({ title: req.body.title });
-    res.redirect("/");
-});
+router.post("/add-product", productController.postAddProduct);
 
 // /admin/add-product => GET
-router.get("/add-product", (req, res, next) => {
-    res.render("add-product.ejs", {
-        pageTitle: "Add Product",
-        path: "/admin/add-product",
-        activeAddProduct: true,
-        addProductCSS: true,
-        shopCSS: true,
-    });
-});
+router.get("/add-product", productController.getAddProduct);
 
-module.exports.routes = router;
-module.exports.products = products;
+module.exports = router;
